static_cast<int>

+++ hungarian.cpp
@@ -828,12 +828,13 @@ int CoverZeros::find_row_without_choice(size_t choice_column_index)
 // 返回类型不能是size_t的原因是，需要用-1来验证错误
 {
     int row_without_choice_index = -1;
-    for(size_t i = 0; i < zero_locations.size(); ++i){ // 遍历每一行
+    for(size_t i = 0; i < zero_locations.size(); ++i)
+    { // 遍历每一行
 
         // 判断在指定列是否存在0，并且这行里没有选择
         if(zero_locations[i][choice_column_index] && std::none_of(choices[i].begin(), choices[i].end(), [](bool b){ return b; }))
         {
-            row_without_choice_index = i;
+            row_without_choice_index = static_cast<int>(i);
 
             break;
         }
